requirements-dev.txt
-r requirements.txt
pytest==8.3.4
pytest-asyncio==0.24.0
fakeredis[lua]==2.26.2
hookify @ git+https://github.com/kiurakku/Hookify.git
ruff==0.8.4
